Support on balance for case of providers which do not exist, removed manual temporary directory selection to instead use temp file found in _server

+++ PluginAthenaHealth/DataContracts/DataWrappers.cs
@@ -122,7 +122,7 @@ namespace PluginAthenaHealth.DataContracts
     public class Provider
     {
         [JsonProperty("providerid")]
-        public int ProviderId { get; set; }
+        public long ProviderId { get; set; }
     }
 
     public class PatientAppointmentReasons
